- Tentative version of ThreadPool -
Custom thread objects which allow for a better composition and allow some sort of scheduling. NOTE: Thread objects have 128 bytes in size (8 from thread, 40 from mutex and 80 from queue) could be brought down to 72 at minimum. This could be harmful for caching and will have to be tested. Also getting the task queue size for the scheduler is hacky and would otherwise require all mutexes to be locked every time which would be insane and create a lot of contention. This will have to be compared against a ThreadPool class that owns everything.